requirements.txt
scikit-learn
pandas
matplotlib
scipy
jupyter
django==2.0.5
django-rest-swagger
djangorestframework==3.8.2
httpie
psycopg2
